refactor(pool): inline OpenedVirtualDoc into document_tracker
Consolidates the 17-line virtual_doc.rs module into document_tracker.rs
to reduce module fragmentation and make the tight coupling between
OpenedVirtualDoc and DocumentTracker more explicit.
